schema wired

+++ data/migrations/01-make_cars_table.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
     tbl.string('vin', 13).notNullable().unique();
     tbl.string('make', 100).notNullable();
     tbl.string('model', 100).notNullable();
-    tbl.numeric('mileage').unsigned().notNullable();
+    tbl.integer('mileage').unsigned().notNullable();
     tbl.string('title', 100).defaultTo('clean');
     tbl.string('transmission').defaultTo('automatic');
   });
